requirements.txt
selenium==3.141.0
webdriver_manager==3.4.2
spotipy
python-dotenv
